requirements.txt
google-generativeai
python-dotenv
PyQt6